feat : Text size in double

+++ Server/src/main/java/fr/nathan/mim/render/shape/shapes/Text.java
@@ -9,12 +9,12 @@ import lombok.ToString;
 @ToString(callSuper = true)
 public class Text extends Shape {
 
-    private final int size;
+    private final double size;
     private final Point2D center;
     private final String value;
     private final double radians;
 
-    public Text(Point2D center, String value, int size, double radians) {
+    public Text(Point2D center, String value, double size, double radians) {
         this.center  = center;
         this.value   = value;
         this.size    = size;
